This branch will explore indirect btrees; the primary motivation is assess handling of variable length data by simply copying and reworking the existing fixed length data approach of boost/btree/set.hpp and map.hpp.
This is a follow on to the "vary" branch, which isn't going well, presumably because I don't know enough about a varying length implementation to abstract it. So this branch will just duplicate code and then rework it for variable length data. If it works, I can then study the difference from master to get a better idea of how to abstract away those differences.